Moved load ecqm-content script to dbSetup.ts
