+++ service/.gitignore
@@ -4,3 +4,4 @@ node_modules
 .DS_Store
 cache
 ecqm-content-r4-2021
+ecqm-content-qicore-2024
